+++ cricket_manager/ui/widgets/__init__.py
@@ -10,6 +10,7 @@ from .modal import Modal
 from .over_beads import OverBeads
 from .radar_chart import RadarChart
 from .slider import Slider
+from .tab_bar import TabBar
 from .star_rating import StarRating
 from .shot_map import ShotMap
 from .bowling_map import BowlingMap
@@ -19,6 +20,6 @@ from .country_flag import draw_country_flag
 
 __all__ = [
     "AttributeBar", "Button", "ButtonStyle", "Card", "ComparisonPanel",
-    "DataTable", "FormGraph", "Modal", "OverBeads", "RadarChart", "Slider", "StarRating", "ShotMap", "BowlingMap",
+    "DataTable", "FormGraph", "Modal", "OverBeads", "RadarChart", "Slider", "StarRating", "TabBar", "ShotMap", "BowlingMap",
     "WeatherDisplay", "PitchDisplay", "draw_country_flag",
 ]